// .pnpm-store/v11/projects/52bc1ab9110c2017903fef4e203abe4c/src/types/apiTypes.ts
export interface ModuleParameter {
  id: number;
  name: string;
  code: string;
  type: 'string' | 'number' | 'boolean' | 'date' | 'select' | 'image';
  required: boolean;
  value?: string | number | boolean | null;
  options?: string[]; // Pour les champs de type select
}

export interface ModuleAssignment {
  id: number;
  price?: number | null;
  module: {
    id: number;
    name: string;
    label?: string;
    description?: string | null;
    icon: string | null;
    category: string | null;
    status: string;
    price?: number | null;
    createdAt: string;
  };
  parameters: ModuleParameter[];
  actif: boolean;
  ordre: number;
}